Initialize clusters after pretraining.
Cluster initialization is now performed after the pretraining step when
the mode is set to pretrain.

Cluster initialization is now only performed at the beginning of the
full training step with a supplied pretrained model when
--train_init_clusters is true.

An additional mode that only performs cluster initialization has been
added and can be activated using --mode init_clusters.