Add support for SSIM metrics.
ssim_matrix.py can be used to generate an ssim matrix from all of the
images in a directory (including images in subdirectores). This script
will also create a directory with symlinks to the images in the SSIM
matrix, with one image symlink per subdirectory, so that the dataset
loader for the custom dataset in torch_DCEC.py will label each image
with its index in the SSIM matrix.

Once the ssim matrix has been generated, it can be used with
torch_DCEC.py by using the --ssim_matrix argument. (Note that setting
the dataset path to point to the directory of symlinks generated with
the ssim matrix is required).